Файл main.js разбит на модули
