+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кекстаграм»
 
 * Студент: [Yana Semenova](https://up.htmlacademy.ru/javascript/31/user/1040325).
-* Наставник: `Сергей Рожков`.
+* Наставник: `Олег Глущенко`.
 
 ---
