+++ js/constants.js
@@ -1,0 +1,58 @@
+const descriptions = [
+  'Эт я тут отдыхал',
+  'Все на пляж!',
+  'Цвет воды просто потрясный',
+  'Это я!',
+  'Зацените что приготовил',
+  'Такую машинку хочу',
+  'На диете, на завтрак одна клубничинка',
+  'Хотите рецепт?',
+  'Так низко летит!',
+  'Крутой органайзер',
+  'Очень солнечно сегодня!',
+  'Вчера купил',
+  'На пикнике',
+  'Лучшая сушинка',
+  'Модные тапки',
+  'На отдых летел',
+  'Был на концерте',
+  'Всегда такую хотел',
+  'Не знаю зачем, но пусть будут',
+  'Отдыхаю',
+  'На пп с сегодняшнего дня!',
+  'Красивый закат',
+  'Модный крабик',
+  'Концерт был супер',
+  'Приключение огонь',
+];
+
+const names = [
+  'Вася',
+  'Петя',
+  'Конфетка',
+  'Маша',
+  'Алиса',
+  'Картофельная долька',
+  'Анастасия',
+  'Елена',
+  'Персевальд',
+  'Персик',
+  'Оксана',
+  'Бусинка',
+  'Пушинка',
+  'Булочка',
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо.',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+export { descriptions, names, messages };
