+++ package.json
@@ -3,6 +3,7 @@
   "version": "30.0.0",
   "private": true,
   "description": "Личный проект «Кекстаграм» от HTML Academy",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git@github.com:htmlacademy-javascript/1040325-kekstagram-31.git"
